osint_tool/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
rich==13.6.0
click==8.1.7
colorama==0.4.6
shodan==1.30.1
virustotal-api==1.1.11
sherlock-project==0.15.0
dnspython==2.4.2
python-whois==0.8.0
Pillow==10.1.0
exifread==3.0.0
beautifulsoup4==4.12.2
pyyaml==6.0.1
pyperclip==1.8.2
cryptography==41.0.4 